+++ src/lib/output/get-runtime-import.js
@@ -2,7 +2,7 @@ import path from 'path';
 
 import {stringifyRequest} from 'loader-utils';
 
-import {getImportStr} from '../../utils/get-import-str';
+import {getImportStr} from '../utils/get-import-str';
 
 export function getRuntimeImport(loaderContext, esModule) {
     const runtimePath = stringifyRequest(
