+++ src/lib/utils/get-first-existed-path.js
