+++ src/lib/utils/get-first-existed-path.spec.js
@@ -1,4 +1,4 @@
-import {ImportWrapper} from '../lib/import-wrapper/ImportWrapper';
+import {ImportWrapper} from '../import-wrapper/ImportWrapper';
 
 import {getFirstExistedPath} from './get-first-existed-path';
