+++ src/lib/utils/get-path-remover.js
