+++ src/lib/utils/get-possible-paths.js
