+++ src/lib/utils/is-unique-object.js
