+++ src/lib/utils/to-regexp-source.js
